# apps/worker/docs/swagger.yaml
definitions:
...
  internal_handler.changePasswordRequest:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
  internal_handler.changeUsernameRequest:
    properties:
      username:
        type: string
    type: object
...
  internal_handler.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: LLM API Gateway — aggregate providers, manage load balancing, and track
    usage/costs.
  title: Wheel API
  version: 1.0.0
paths:
...
  /api/v1/user/change-password:
    post:
      consumes:
        - application/json
      parameters:
        - description: Old and new password
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/internal_handler.changePasswordRequest"
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
        "401":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Change admin password
      tags:
        - Auth
  /api/v1/user/change-username:
    post:
      consumes:
        - application/json
      parameters:
        - description: New username
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/internal_handler.changeUsernameRequest"
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Change admin username
      tags:
        - Auth
  /api/v1/user/login:
    post:
      consumes:
        - application/json
      parameters:
        - description: Login credentials
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/internal_handler.loginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {token, expireAt}}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
        "401":
          description: "{success: false, error: string}"
          schema:
            type: object
      summary: Admin login
      tags:
        - Auth
  /api/v1/user/status:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {authenticated: true}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Check authentication status
      tags:
        - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"